requirements.txt
jsonschema == 4.21.1
langchain==0.1.13
langdetect==1.0.9

pydantic==2.6.4
spacy==3.7.4
transformers==4.39.0
git+https://github.com/boudinfl/pke.git
torch==2.2.1
torchvision==0.17.1
torchaudio==2.2.1
sentencepiece==0.2.0


streamlit==1.32.2
pypdf==4.1.0
python-docx ==1.1.0
ebisu==2.1.0